[update] new outline
