+++ fedlab/core/client/trainer.py
@@ -17,8 +17,10 @@ from typing import List
 
 import torch
 
-from ..client import ORDINARY_TRAINER
-from ..model_maintainer import ModelMaintainer
+from fedlab.dataset.dataset import FedLabDataset
+
+from ..client import ORDINARY_TRAINER, SERIAL_TRAINER
+from ..model_maintainer import ModelMaintainer, SerialModelMaintainer
 from ...utils import Logger, SerializationTool
 
 
@@ -39,6 +41,7 @@ class ClientTrainer(ModelMaintainer):
     def __init__(self, model, cuda):
         super().__init__(model, cuda)
         self.client_num = 1  # default is 1.
+        self.dataset = FedLabDataset()
         self.type = ORDINARY_TRAINER
 
     @abstractproperty
@@ -66,71 +69,73 @@ class ClientTrainer(ModelMaintainer):
         """Evaluate quality of local model."""
         raise NotImplementedError()
 
-
-class SGDClientTrainer(ClientTrainer):
-    """Client backend handler, this class provides data process method to upper layer.
+class SerialClientTrainer(SerialModelMaintainer):
+    """Base class. Simulate multiple clients in sequence in a single process.
 
     Args:
-        model (torch.nn.Module): PyTorch model.
-        data_loader (torch.utils.data.DataLoader): :class:`torch.utils.data.DataLoader` for this client.
-        epochs (int): the number of local epoch.
-        optimizer (torch.optim.Optimizer): optimizer for this client's model.
-        criterion (torch.nn.Loss): loss function used in local training process.
-        cuda (bool, optional): use GPUs or not. Default: ``False``.
-        logger (Logger, optional): :object of :class:`Logger`.
+        model (torch.nn.Module): Model used in this federation.
+        client_num (int): Number of clients in current trainer.
+        cuda (bool): Use GPUs or not. Default: ``False``.
+        logger (Logger, optional): Object of :class:`Logger`.
     """
 
-    def __init__(self,
-                 model,
-                 data_loader,
-                 epochs,
-                 optimizer,
-                 criterion,
-                 cuda=False,
-                 logger=None):
-        super(SGDClientTrainer, self).__init__(model, cuda)
-
-        self._data_loader = data_loader
-        self.epochs = epochs
-        self.optimizer = optimizer
-        self.criterion = criterion
+    def __init__(self, model, client_num, cuda=False, logger=None):
+        super().__init__(model, cuda)
+        self.client_num = client_num
+        self.dataset = FedLabDataset()
+        self.type = SERIAL_TRAINER  # represent serial trainer
         self._LOGGER = Logger() if logger is None else logger
-
-        self.model_time = 0
+        self.param_list = []
 
     @property
     def uplink_package(self):
-        """Return a tensor list for uploading to server.
+        return self.param_list
 
-            This attribute will be called by client manager.
-            Customize it for new algorithms.
+    # def _train_alone(self, model_parameters, train_loader):
+    #     """Train local model with :attr:`model_parameters` on :attr:`train_loader`.
+        
+    #     Args:
+    #         model_parameters (torch.Tensor): Serialized model parameters of one model.
+    #         train_loader (torch.utils.data.DataLoader): :class:`torch.utils.data.DataLoader` for this client.
+    #     """
+    #     raise NotImplementedError()
+
+    # def _get_dataloader(self, client_id):
+    #     """Get :class:`DataLoader` for ``client_id``."""
+    #     raise NotImplementedError()
+
+    @abstractclassmethod
+    def local_process(self, payload) -> bool:
+        """Manager of the upper layer will call this function with accepted payload
+        
+            In synchronous mode, return True to end current FL round.
         """
-        return [self.model_parameters]
+        raise NotImplementedError()
 
-    def local_process(self, payload):
-        model_parameters = payload[0]
-        self.train(model_parameters)
+    def train(self):
+        """Override this method to define the algorithm of training your model. This function should manipulate :attr:`self._model`"""
+        raise NotImplementedError()
 
-    def train(self, model_parameters) -> None:
-        """Client trains its local model on local dataset.
+    def evaluate(self):
+        """Evaluate quality of local model."""
+        raise NotImplementedError()
 
-        Args:
-            model_parameters (torch.Tensor): Serialized model parameters.
-        """
-        SerializationTool.deserialize_model(
-            self._model, model_parameters)  # load parameters
-        self._LOGGER.info("Local train procedure is running")
-        for ep in range(self.epochs):
-            self._model.train()
-            for inputs, labels in self._data_loader:
-                if self.cuda:
-                    inputs, labels = inputs.cuda(self.gpu), labels.cuda(
-                        self.gpu)
+    # def local_process(self, id_list, payload):
+    #     """Train local model with different dataset according to client id in ``id_list``.
 
-                outputs = self._model(inputs)
-                loss = self.criterion(outputs, labels)
+    #     Args:
+    #         id_list (list[int]): Client id in this training serial.
+    #         payload (list[torch.Tensor]): communication payload from server.
+    #     """
+    #     self.buffer = []
+    #     self._LOGGER.info(
+    #         "Local training with client id list: {}".format(id_list))
+    #     for idx in id_list:
+    #         self._LOGGER.info(
+    #             "Starting training procedure of client [{}]".format(idx))
 
-                self.optimizer.zero_grad()
-                loss.backward()
-                self.optimizer.step()
-        self._LOGGER.info("Local train procedure is finished")
+    #         # data_loader = self._get_dataloader(client_id=idx)
+    #         # self._train_alone(model_parameters=payload[0],
+    #         #                   train_loader=data_loader)
+    #         self.buffer.append(self.model_parameters)
+    #     return self.buffer
